third_party/blink/web_tests/platform/generic/svg/as-image/smil-usecounter-in-image-expected.txt

PASS internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is false
PASS internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache) is true
PASS successfullyParsed is true

TEST COMPLETE
